+++ js/edit-image.js
@@ -1,0 +1,1 @@
+// Функциональность по  редактированию изображения
