+++ js/filter-image.js
@@ -1,0 +1,1 @@
+// Функциональсность по фильтрации изображений от других пользователей
