+++ js/form.js
@@ -1,0 +1,1 @@
+// Функциональность по отправке данных на сервер
