+++ js/preview-image.js
@@ -1,0 +1,1 @@
+// Функциональность по просмотру загруженных изображений
