+++ js/upload-new-image.js
@@ -1,0 +1,1 @@
+// Функциональность по загрузке нового изображения
